Adding fix for incorrect log level bug with refactoring entrypoint.py to fix circular dependency issue
